Added Query Builder Partition modules
