+++ lib/index.js
@@ -1,9 +1,21 @@
 'use strict';
 
 const QueryBuilder = require('./query-builder');
+const QueryBuilderFields = require('./query-builder-fields');
+const QueryBuilderJoins = require('./query-builder-joins');
+const QueryBuilderFilters = require('./query-builder-filters');
+const QueryBuilderGroup = require('./query-builder-group');
+const QueryBuilderOrder = require('./query-builder-order');
+const QueryBuilderPagination = require('./query-builder-pagination');
 const QueryBuilderError = require('./query-builder-error');
 
 module.exports = {
 	QueryBuilder,
-	QueryBuilderError
+	QueryBuilderError,
+	QueryBuilderFields,
+	QueryBuilderFilters,
+	QueryBuilderGroup,
+	QueryBuilderJoins,
+	QueryBuilderOrder,
+	QueryBuilderPagination
 };
