+++ lib/query-builder-error.js
@@ -20,7 +20,7 @@ class QueryBuilderError extends Error {
 			NO_VALUES: 13,
 			NOTHING_SELECT: 14,
 			DUPLICATE_ITEM: 15,
-			MYSQL_DRIVER_ERROR: 16
+			DRIVER_ERROR: 16
 
 		};
